backend/requirements.txt
fastapi==0.131.0
uvicorn[standard]==0.41.0

langchain==0.1.17
langchain-community==0.0.36
langchain-openai==0.1.7

chromadb==0.4.24
pypdf==6.7.2

openai==1.30.1
httpx==0.27.0
python-dotenv==1.2.1